AESP-Microservice-Project/services/AI-core-service/requirements.txt
flask
flask-cors
flask-sqlalchemy
pymysql
groq
python-dotenv
flask-jwt-extended
requests
cryptography